+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- Improved documentation of a project creation in the QT creator
+- Added `.gitignore` file generation
+
 ## [0.2.0] - 2018-06-17
 ### Added
 - Added documentation about *Qt Creator* usage.